requirements.txt
google-cloud-asset==3.17.1
pandas==1.5.3
tabulate==0.9.0
